% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_basic_plots.R
\name{arrange_metagen_plots}
\alias{arrange_metagen_plots}
\title{Arrange the forest plots created by create_metagen_plots on a grid}
\usage{
arrange_metagen_plots(plots = list(), col_count = 1)
}
\arguments{
\item{plots}{a list of forest plots}

\item{col_count}{the number of columns to plot}
}
\description{
Arrange the forest plots created by create_metagen_plots on a grid
}
