% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_basic_plots.R
\name{calculate_metagen_ticks}
\alias{calculate_metagen_ticks}
\title{Calculate the x-tick positions for a forest plot. This is usually unneeded,
but sometimes the plots don't produce ticks on their own}
\usage{
calculate_metagen_ticks(df_plot = data.frame())
}
\arguments{
\item{df_plot}{as data.frame of the plotting data}
}
\description{
Calculate the x-tick positions for a forest plot. This is usually unneeded,
but sometimes the plots don't produce ticks on their own
}
