% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting_tables.R
\name{checkbox_js}
\alias{checkbox_js}
\title{create the javascript code for a checkbox-button in a data frame so that it
returns information when selected}
\usage{
checkbox_js(
  dtid = data.frame(),
  ns = shiny::NS(NULL),
  column = 1,
  name = "checkb"
)
}
\arguments{
\item{dtid}{the dataframe id used as an identifier}

\item{ns}{The NS function to add the shiny module id to the label}

\item{column}{the column number to return when selected}

\item{name}{the name identifier to use for the button}
}
\description{
create the javascript code for a checkbox-button in a data frame so that it
returns information when selected
}
