% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{convert_hyperlink_link}
\alias{convert_hyperlink_link}
\title{Convert a Shiny datatable hyper link into a url}
\usage{
convert_hyperlink_link(hyper_link = character())
}
\arguments{
\item{hyper_link}{the datatable hyper link to convert to a url}
}
\description{
Convert a Shiny datatable hyper link into a url
}
