% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_convert_variables.R
\name{convert_range_to_list}
\alias{convert_range_to_list}
\title{convert a genomic range from CHR:START-END format to the actual values}
\usage{
convert_range_to_list(genomic_range = character())
}
\arguments{
\item{genomic_range}{a string in the format CHR:START-END}
}
\description{
convert a genomic range from CHR:START-END format to the actual values
}
