% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gviz_tracks.R
\name{convert_track_colors_to_list}
\alias{convert_track_colors_to_list}
\title{Some functions require track colors to be a named list instead of a
data.frame. This function runs that conversion}
\usage{
convert_track_colors_to_list(df_color, label_col = "label")
}
\arguments{
\item{df_color}{a data.frame of the labels and track colors}

\item{label_col}{the column that has the color labels}
}
\description{
Some functions require track colors to be a named list instead of a
data.frame. This function runs that conversion
}
