% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{create_Agora_link}
\alias{create_Agora_link}
\title{Create links tot he Agora database for genes}
\usage{
create_Agora_link(gene = character(), df_ens = data.frame())
}
\arguments{
\item{gene}{a vector of genes to convert to Agora links}

\item{df_ens}{a dataframe of genes present in Agora, and their ENS IDs}
}
\description{
Create links tot he Agora database for genes
}
