% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gviz_tracks.R
\name{create_empty_annotation_track}
\alias{create_empty_annotation_track}
\title{Create an empty annotation track. This can be run if there is no data over
the region of interest, as the standard code would throw an error in that
case}
\usage{
create_empty_annotation_track(chromosome = "chr1", rstart = 0, rend = 0)
}
\arguments{
\item{chromosome}{The chromosome where the genomic region is located}

\item{rstart}{The start of the genomic region}

\item{rend}{The end of the genomic region}
}
\description{
Create an empty annotation track. This can be run if there is no data over
the region of interest, as the standard code would throw an error in that
case
}
