% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting_tables.R
\name{create_empty_dataframe}
\alias{create_empty_dataframe}
\title{Initialize an empty dataframe to store dataset plotting values in}
\usage{
create_empty_dataframe(source = FALSE, metric = FALSE)
}
\arguments{
\item{source}{a boolean whether to store the dataset source}

\item{metric}{a boolean about whether to have a p-value cutoff (True), or a checkbox on whether it can be selected (FALSE)}
}
\description{
Initialize an empty dataframe to store dataset plotting values in
}
