% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{create_empty_statistics_table}
\alias{create_empty_statistics_table}
\title{Create an empty data frame to store statistics, with all relevant columns
defined. Columns and their variable types are defined to avoid potential
errors downstream}
\usage{
create_empty_statistics_table(table_category = "CpG")
}
\arguments{
\item{table_category}{whether the output should be for "CpG", "Gene" or "Genomewide" tabs}
}
\description{
Create an empty data frame to store statistics, with all relevant columns
defined. Columns and their variable types are defined to avoid potential
errors downstream
}
