% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gviz_tracks.R
\name{create_ENSEMBL_track}
\alias{create_ENSEMBL_track}
\title{Create a Gviz track from ENSEMBL data over the genomic region}
\usage{
create_ENSEMBL_track(
  fname,
  label = "ENSEMBL Genes",
  genome = "hg19",
  chromosome = "chr1",
  rstart = 0,
  rend = 0
)
}
\arguments{
\item{fname}{The file path where the track is stored}

\item{label}{Identifier whether to plot genes (ENSEMBL Genes) or transcripts (ENSEMBL Transcripts)}

\item{genome}{whether to use the hg19 or hg38 human genomes}

\item{chromosome}{The chromosome where the genomic region is located}

\item{rstart}{The start of the genomic region}

\item{rend}{The end of the genomic region}
}
\description{
Create a Gviz track from ENSEMBL data over the genomic region
}
