% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting_tables.R
\name{create_epigenetic_empty_dataframe}
\alias{create_epigenetic_empty_dataframe}
\title{Initialize an empty dataframe to store epigenetic data in}
\usage{
create_epigenetic_empty_dataframe()
}
\description{
Initialize an empty dataframe to store epigenetic data in
}
