% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_ui.R
\name{create_gene_conditional_input}
\alias{create_gene_conditional_input}
\title{Create the gene or genomic-range selector input}
\usage{
create_gene_conditional_input(ns = shiny::NS(NULL), id = "")
}
\arguments{
\item{ns}{Shiny namespace}

\item{id}{Shiny namespace id}
}
\description{
Create the gene or genomic-range selector input
}
