% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_ui.R
\name{create_genome_version_input}
\alias{create_genome_version_input}
\title{Create the genome version selector input}
\usage{
create_genome_version_input(ns = shiny::NS(NULL), tour = TRUE)
}
\arguments{
\item{ns}{Shiny namespace}

\item{tour}{Whether or not to wrap a tour container around the input}
}
\description{
Create the genome version selector input
}
