% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{create_GWAS_gene_link}
\alias{create_GWAS_gene_link}
\title{Create links to the GWAS gene inforamtion}
\usage{
create_GWAS_gene_link(genes = character())
}
\arguments{
\item{genes}{genes to convert to GWAS links}
}
\description{
Create links to the GWAS gene inforamtion
}
