% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{create_GWAS_region_link}
\alias{create_GWAS_region_link}
\title{Create links to the GWAS genomic range inforamtion}
\usage{
create_GWAS_region_link(ranges = character())
}
\arguments{
\item{ranges}{genomic ranges to convert to GWAS links}
}
\description{
Create links to the GWAS genomic range inforamtion
}
