% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{create_London_Link}
\alias{create_London_Link}
\title{Create links to the epigenetic essex blood-brain database for cpgs present there}
\usage{
create_London_Link(cpg = character(), cpg_London = character(), excel = FALSE)
}
\arguments{
\item{cpg}{a vector of cpgs to convert to blood-brain links}

\item{cpg_London}{a vector of cpgs present in the blood-brain database}

\item{excel}{if true, will create a simply hyperlink that can be put in an excel file. Otherwise, it will create a tag that can be put in a Shiny datatable}
}
\description{
Create links to the epigenetic essex blood-brain database for cpgs present there
}
