% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_basic_plots.R
\name{create_metagen_plots}
\alias{create_metagen_plots}
\title{Create a list of forest plots using the meta package}
\usage{
create_metagen_plots(df_plot_forest = data.frame())
}
\arguments{
\item{df_plot_forest}{a data frame containing the meta information for each plot}
}
\description{
Create a list of forest plots using the meta package
}
