% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{create_mQTL_Link}
\alias{create_mQTL_Link}
\title{Create links to the mQTL database for cpgs present there}
\usage{
create_mQTL_Link(cpg = character(), cpg_mQTL = character(), excel = FALSE)
}
\arguments{
\item{cpg}{a vector of cpgs to convert to mQTL links}

\item{cpg_mQTL}{a vector of cpgs present in the mQTL database}

\item{excel}{if true, will create a simply hyperlink that can be put in an excel file. Otherwise, it will create a tag that can be put in a Shiny datatable}
}
\description{
Create links to the mQTL database for cpgs present there
}
