% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{create_Niagads_link}
\alias{create_Niagads_link}
\title{Create links to the Niagads database for genes}
\usage{
create_Niagads_link(gene = character(), df_ens = data.frame())
}
\arguments{
\item{gene}{a vector of genes to convert to Niagads links}

\item{df_ens}{a dataframe of genes present in Niagads, and their ENS IDs}
}
\description{
Create links to the Niagads database for genes
}
