% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{create_statistics_table}
\alias{create_statistics_table}
\title{Read in CpG statistics, and convert them into a statistics table used to
display datatables and for downloads}
\usage{
create_statistics_table(
  df_stats = data.frame(),
  df_labels = data.frame(),
  df_cpg_features = data.frame(),
  table_category = "CpG"
)
}
\arguments{
\item{df_stats}{a data frame of the CpG statistics}

\item{df_labels}{a data frame with metadata information about the datasets}

\item{df_cpg_features}{a data frame of CpG positional information}

\item{table_category}{whether the output should be for "CpG", "Gene" or "Genomewide" tabs}
}
\description{
Read in CpG statistics, and convert them into a statistics table used to
display datatables and for downloads
}
