% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gviz_tracks.R
\name{create_track_legend}
\alias{create_track_legend}
\title{Create a colors legend for chromHMM roadmap or chromHMM track}
\usage{
create_track_legend(track)
}
\arguments{
\item{track}{the category to create a legend for. Either 'Roadmap Legend' or 'ChromHMM Legend'}
}
\description{
Create a colors legend for chromHMM roadmap or chromHMM track
}
