% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gviz_tracks.R
\name{create_txdb_track}
\alias{create_txdb_track}
\title{Create a track of UCSC gene data}
\usage{
create_txdb_track(genome = "hg19", chromosome = "chr1", rstart = 0, rend = 0)
}
\arguments{
\item{genome}{whether to use the hg19 or hg38 human genomes}

\item{chromosome}{The chromosome where the genomic region is located}

\item{rstart}{The start of the genomic region}

\item{rend}{The end of the genomic region}
}
\description{
Create a track of UCSC gene data
}
