% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{filter_cpg_sql_statistics}
\alias{filter_cpg_sql_statistics}
\title{scan for cpgs in datasets by significance thresholds}
\usage{
filter_cpg_sql_statistics(
  datasets = character(),
  sources = character(),
  thresholds = numeric(),
  filters = character(),
  metrics = character()
)
}
\arguments{
\item{datasets}{a list of datasets to search in}

\item{sources}{a list of sources corresponding to the datasets}

\item{thresholds}{a list of significant thresholds}

\item{filters}{a list of filters whether to look for values above (\code{>}) or below (\code{<}) the thresholds}

\item{metrics}{a list of significance metrics (pValue or FDR)}
}
\description{
scan for cpgs in datasets by significance thresholds
}
