% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_basic_plots.R
\name{fix_venn_label_positions}
\alias{fix_venn_label_positions}
\title{Adjust venn diagram label positions so that they don't go off the page as often}
\usage{
fix_venn_label_positions(p, label_count = 2, method = "epigenetic")
}
\arguments{
\item{p}{the ggplot plot data}

\item{label_count}{the number of datasets that are on the venn diagram}

\item{method}{whether the data came from the epigenetic or genome-wide tab}
}
\description{
Adjust venn diagram label positions so that they don't go off the page as often
}
