% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_basic_plots.R
\name{format_gene_manhattan_data}
\alias{format_gene_manhattan_data}
\title{filter data for a manhattan plot, and put it into the right format}
\usage{
format_gene_manhattan_data(df_gene_manplot = data.frame())
}
\arguments{
\item{df_gene_manplot}{a data frame containing the plotting information}
}
\description{
filter data for a manhattan plot, and put it into the right format
}
