% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{format_pvalues_column}
\alias{format_pvalues_column}
\title{Format p-values so that they are in scientific notation, with 3 digits}
\usage{
format_pvalues_column(pvalues = numeric())
}
\arguments{
\item{pvalues}{a vector of p-values}
}
\description{
Format p-values so that they are in scientific notation, with 3 digits
}
