% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gviz_tracks.R
\name{generate_track}
\alias{generate_track}
\title{Create a Gviz track over the region of interest.}
\usage{
generate_track(
  category = "Roadmap",
  genome = "hg19",
  fname = character(),
  label = character(),
  chromosome = "chr1",
  rstart = 0,
  rend = 0
)
}
\arguments{
\item{category}{the type of Gviz track to create}

\item{genome}{whether to use the hg19 or hg38 human genomes}

\item{fname}{The file path where the track is stored}

\item{label}{An extra identifier for the subtype within the category}

\item{chromosome}{The chromosome where the genomic region is located}

\item{rstart}{The start of the genomic region}

\item{rend}{The end of the genomic region}
}
\description{
Create a Gviz track over the region of interest.
}
