% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gviz_tracks.R
\name{generate_track_plots}
\alias{generate_track_plots}
\title{Convert a series of Gviz tracks into a single plot with a title over them}
\usage{
generate_track_plots(
  plot_tracks = list(),
  titles = list(),
  chromosome = "chr1",
  start = 0,
  end = 0
)
}
\arguments{
\item{plot_tracks}{a list of Gviz trakcs}

\item{titles}{a list of titles to put above the Gviz tracks. If these can't be supplied, then using Gviz::plotTracks is more convenient}

\item{chromosome}{the chromosome that the plot is on}

\item{start}{the start position of the tracks to plot}

\item{end}{the end position of the tracks to plot}
}
\description{
Convert a series of Gviz tracks into a single plot with a title over them
}
