% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{get_cpg_sql_parameters}
\alias{get_cpg_sql_parameters}
\title{Get the gene associations for a given list of cpgs}
\usage{
get_cpg_sql_parameters(cpgs = character(), unique = TRUE)
}
\arguments{
\item{cpgs}{a list of cpgs of interest}

\item{unique}{a boolean of whether we want only uniquely associate genes, or if repeats are allowed}
}
\description{
Get the gene associations for a given list of cpgs
}
