% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{get_cpg_sql_positions}
\alias{get_cpg_sql_positions}
\title{get cpg position statistics in a given genome version}
\usage{
get_cpg_sql_positions(cpgs = character(), genome_version = "hg19")
}
\arguments{
\item{cpgs}{a list of cpgs of interest}

\item{genome_version}{the version of the genome to search. Either hg19 or hg38}
}
\description{
get cpg position statistics in a given genome version
}
