% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{get_cpg_sql_statistics}
\alias{get_cpg_sql_statistics}
\title{Get the statistics for a given list of cpgs in a given list of datasets}
\usage{
get_cpg_sql_statistics(
  cpgs = character(),
  datasets = character(),
  sources = character()
)
}
\arguments{
\item{cpgs}{a list of cpgs of interest}

\item{datasets}{a list of datasets where we are interested in the cpgs}

\item{sources}{either empty, or a list of sources corresponding to each dataset}
}
\description{
Get the statistics for a given list of cpgs in a given list of datasets
}
