% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{get_gene_sql_genome_list}
\alias{get_gene_sql_genome_list}
\title{scan for a list of all genes in a genome}
\usage{
get_gene_sql_genome_list(genome_version = "hg19")
}
\arguments{
\item{genome_version}{the version of the genome to search. Either hg19 or hg38}
}
\description{
scan for a list of all genes in a genome
}
