% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{get_gene_sql_positions}
\alias{get_gene_sql_positions}
\title{get gene position statistics in a given genome version}
\usage{
get_gene_sql_positions(gene = character(), genome_version = "hg19")
}
\arguments{
\item{gene}{a gene of interest}

\item{genome_version}{the version of the genome to search. Either hg19 or hg38}
}
\description{
get gene position statistics in a given genome version
}
