% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gviz_tracks.R
\name{get_track_colors}
\alias{get_track_colors}
\title{Read in the color code used for a tarack}
\usage{
get_track_colors(label = character())
}
\arguments{
\item{label}{the label identifying a track. Currently supports 'Roadmap' and 'chromHMM'}
}
\description{
Read in the color code used for a tarack
}
