% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_ui.R
\name{lightButton}
\alias{lightButton}
\title{Create a lighter version of a button}
\usage{
lightButton(...)
}
\arguments{
\item{...}{All parameters passed to shiny actionButton}
}
\description{
This function should only be used if bootstrap 5 is used.
}
