% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_tables.R
\name{merge_cpg_statistical_data}
\alias{merge_cpg_statistical_data}
\title{Merge CpG statics, and dataset metadata tables into 1 dataframe for
display and plotting}
\usage{
merge_cpg_statistical_data(
  df_stats = data.frame(),
  df_labels = data.frame(),
  df_cpg_features = data.frame()
)
}
\arguments{
\item{df_stats}{a data frame of the CpG statistics}

\item{df_labels}{a data frame with metadata information about the datasets}

\item{df_cpg_features}{a data frame of CpG positional information}
}
\description{
Merge CpG statics, and dataset metadata tables into 1 dataframe for
display and plotting
}
