% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_basic_plots.R
\name{plot_dataset_metagen}
\alias{plot_dataset_metagen}
\title{Create a forest plot using the meta package}
\usage{
plot_dataset_metagen(df_plot = data.frame(), title = "")
}
\arguments{
\item{df_plot}{a dataframe containing the plotting information}

\item{title}{a string of the title to put over the forest plot}
}
\description{
Create a forest plot using the meta package
}
