% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_basic_plots.R
\name{plot_gene_manhattan_data}
\alias{plot_gene_manhattan_data}
\title{Create a manhattan plot of gene CpG information}
\usage{
plot_gene_manhattan_data(
  df_gene_manplot = data.frame(),
  df_gene_label = data.frame(),
  start = 0,
  end = 0,
  threshold = 0
)
}
\arguments{
\item{df_gene_manplot}{a data frame containing cpg positional and p-value inofrmation}

\item{df_gene_label}{a data frame containing cpg information on cpgs to label}

\item{start}{an integer for the x-axis left-most position}

\item{end}{an integer for the x-axis right-most position}

\item{threshold}{the threshold for the -log10(p-value) to label with a horizontal line}
}
\description{
Create a manhattan plot of gene CpG information
}
