% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_basic_plots.R
\name{plot_list_metagen}
\alias{plot_list_metagen}
\title{Generate a list of forest plots and arrange them on a grid}
\usage{
plot_list_metagen(df_plot_forest = data.frame(), col_count = 1)
}
\arguments{
\item{df_plot_forest}{a data frame containing the meta information for each plot}

\item{col_count}{the number of columns to plot}
}
\description{
Generate a list of forest plots and arrange them on a grid
}
