% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_ui.R
\name{primaryButton}
\alias{primaryButton}
\title{Create a primary button}
\usage{
primaryButton(...)
}
\arguments{
\item{...}{All parameters passed to shiny actionButton}
}
\description{
Create a primary button
}
