% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gviz_tracks.R
\name{read_and_filter_track}
\alias{read_and_filter_track}
\title{Read in a track and filter to region of interest, and return it as a genomic
range}
\usage{
read_and_filter_track(
  fname = character(),
  genome = "hg19",
  chromosome = "chr1",
  rstart = 0,
  rend = 0
)
}
\arguments{
\item{fname}{file name of track file}

\item{genome}{whether to use the hg19 or hg38 human genomes}

\item{chromosome}{The chromosome where the genomic region is located}

\item{rstart}{The start of the genomic region}

\item{rend}{The end of the genomic region}
}
\description{
Read in a track and filter to region of interest, and return it as a genomic
range
}
