% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{read_in_data}
\alias{read_in_data}
\title{A function to read in all of the data that needs to be stored in memory}
\usage{
read_in_data()
}
\description{
A function to read in all of the data that needs to be stored in memory
}
