% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{scan_cpg_filtered_positions}
\alias{scan_cpg_filtered_positions}
\title{scan for cpgs in a genome by positional parameter}
\usage{
scan_cpg_filtered_positions(
  genome_version = "hg19",
  chromosome = "chr1",
  start = 0,
  end = 0
)
}
\arguments{
\item{genome_version}{the version of the genome to search. Either hg19 or hg38}

\item{chromosome}{the chromosome to search}

\item{start}{the start position of the range of interest}

\item{end}{the end position of the range of interest}
}
\description{
scan for cpgs in a genome by positional parameter
}
