% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{scan_cpg_filtered_statistics}
\alias{scan_cpg_filtered_statistics}
\title{scan the database for cpgs that match a separate significance threshold for each dataset}
\usage{
scan_cpg_filtered_statistics(
  datasets = character(),
  sources = character(),
  thresholds = numeric(),
  filters = character(),
  metrics = character()
)
}
\arguments{
\item{datasets}{a list of datasets}

\item{sources}{a list of sources, the same length as datasets}

\item{thresholds}{a list of significance thresholds, the same length as datasets}

\item{filters}{a list of filters whether to look for values above (\code{>}) or below (\code{<}) the thresholds}

\item{metrics}{a list of significance metrics, the same length as datasets}
}
\description{
scan the database for cpgs that match a separate significance threshold for each dataset
}
