% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{scan_cpg_sql_parameters}
\alias{scan_cpg_sql_parameters}
\title{Get gene relations for a list of cpgs}
\usage{
scan_cpg_sql_parameters(cpgs, unique = TRUE)
}
\arguments{
\item{cpgs}{a list of cpgs of interest}

\item{unique}{a boolean of whether we want only uniquely associate genes, or if repeats are allowed}
}
\description{
Get gene relations for a list of cpgs
}
