% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{scan_sql_query}
\alias{scan_sql_query}
\title{Connect to our SQL database, and get the results of a query}
\usage{
scan_sql_query(query_text = "")
}
\arguments{
\item{query_text}{the SQL query that we will pass}
}
\description{
Connect to our SQL database, and get the results of a query
}
