% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_format_ui.R
\name{twocols}
\alias{twocols}
\title{Modify a checkboxGroupInput to have two columns for the options}
\usage{
twocols(tag)
}
\arguments{
\item{tag}{A checkboxGroupInput tag}
}
\description{
Modify a checkboxGroupInput to have two columns for the options
}
