% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{validate_cpg_search}
\alias{validate_cpg_search}
\title{Run a validation function that checks if there is anything to query for cpgs and datasets}
\usage{
validate_cpg_search(
  cpgs = character(),
  datasets = character(),
  sources = character()
)
}
\arguments{
\item{cpgs}{a list of cpgs that needs to be non-zero length}

\item{datasets}{a list of datasets that needs to be non-zero length}

\item{sources}{a list of dataset sources that needs to be either empty, or as long as datasets, with a 1-to-1 correspondence}
}
\description{
Run a validation function that checks if there is anything to query for cpgs and datasets
}
