% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{validate_threshold_search}
\alias{validate_threshold_search}
\title{Run a validation function that checks if there are an equal number of datasets, sources, thresholds, and metrics}
\usage{
validate_threshold_search(
  datasets = character(),
  sources = character(),
  thresholds = numeric(),
  filters = character(),
  metrics = character()
)
}
\arguments{
\item{datasets}{a list of datasets that need to be non-zero length}

\item{sources}{a list of sources that needs to be the same length as datasets}

\item{thresholds}{a list of significance thresholds that needs to be the same length as datasets}

\item{filters}{a list of filters directions that needs to be the same length as datasets}

\item{metrics}{a list of significance metrics (pValue or FDR) that needs to be the same length as datasets}
}
\description{
Run a validation function that checks if there are an equal number of datasets, sources, thresholds, and metrics
}
