% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_get_data.R
\name{validation_position_inputs}
\alias{validation_position_inputs}
\title{A validation function that checks if there are any cpgs, and we have the correct genome version}
\usage{
validation_position_inputs(cpgs = character(), genome_version = "hg19")
}
\arguments{
\item{cpgs}{a list of cpgs that needs to be non-zero length}

\item{genome_version}{the genome version to search. Needs to be either "hg19" or "hg38"}
}
\description{
A validation function that checks if there are any cpgs, and we have the correct genome version
}
